refactor(firmware): move firmware downloader into carbide-firmware
- move `FirmwareDownloader` out of `carbide-api`
- re-export downloader from `carbide-firmware`
- update API startup and preingestion manager imports
- move downloader tests into firmware crate
