+++ crates/firmware/Cargo.toml
@@ -18,7 +18,11 @@ carbide-api-model = { path = "../api-model", default-features = false }
 
 # these are alphabetized
 eyre = { workspace = true }
+futures-util = { workspace = true }
+md5 = { workspace = true }
+reqwest = { workspace = true }
 serde = { workspace = true }
+tokio = { workspace = true }
 toml = { workspace = true }
 tracing = { workspace = true }
