+++ crates/firmware/src/tests/desired_firmware_versions.rs
@@ -14,6 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+use carbide_macros::sqlx_test;
 use sqlx::FromRow;
 
 use crate::FirmwareConfig;
@@ -23,7 +25,7 @@ struct AsStrings {
     pub versions: String,
 }
 
-#[super::sqlx_test]
+#[sqlx_test]
 pub async fn test_build_versions(pool: sqlx::PgPool) -> Result<(), eyre::Error> {
     use std::ops::DerefMut;
